requirements.txt
auton_survival==0.1.0
keras==2.11.0
lifelines==0.26.4
matplotlib==3.7.4
numpy==1.23.5
openTSNE==1.0.1
pandas==1.5.3
progressbar33==2.4
PyYAML==6.0.1
PyYAML==6.0.1
radiomics==0.1
scikit_learn==1.3.0
scikit_optimize==0.9.0
scipy==1.9.1
seaborn==0.13.2
tensorflow==2.11.0
tensorflow_probability==0.21.0
torch==1.13.1
tqdm==4.66.1
